+++ OguryPresage/OguryPresageAdapter/AppLovinMediationOguryPresageAdapter.h
@@ -1,0 +1,8 @@
+//
+//  AppLovinMediationOguryPresageAdapter.h
+//  AppLovinMediationOguryPresageAdapter
+//
+//  Copyright © 2025 AppLovin. All rights reserved.
+//
+
+#import <AppLovinMediationOguryPresageAdapter/ALOguryPresageMediationAdapter.h>